Refactor vf2pp modules and test files
