--- networkx/algorithms/isomorphism/tests/vf2pp/utils.py
@@ -1,19 +1,0 @@
-import networkx as nx
-
-labels_same = ["blue"]
-
-labels_different = [
-    "white",
-    "red",
-    "blue",
-    "green",
-    "orange",
-    "black",
-    "purple",
-    "yellow",
-    "brown",
-    "cyan",
-    "solarized",
-    "pink",
-    "none",
-]
